upi_fraud_ml/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnspython==2.8.0
joblib==1.5.2
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
pandas==2.3.3
pillow==12.0.0
PyJWT==2.10.1
pymongo==4.15.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyzbar==0.1.9
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.4
threadpoolctl==3.6.0
tzdata==2025.2
